VHR/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy<2.4
opencv-python
open-rppg
